Renaming Form1.cs -> MasterForm.cs & IO.cs -> Utility.cs
