+++ MCC Mod Manager.csproj
@@ -72,18 +72,18 @@
     <Compile Include="Api\Backups.cs" />
     <Compile Include="Api\Config.cs" />
     <Compile Include="Api\Utilities\FolderSelectDialog.cs" />
-    <Compile Include="Form1.cs">
+    <Compile Include="MasterForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form1.Designer.cs">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <Compile Include="MasterForm.Designer.cs">
+      <DependentUpon>MasterForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Api\Utilities\IO.cs" />
+    <Compile Include="Api\Utilities\Utility.cs" />
     <Compile Include="Api\Modpacks.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="Form1.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <EmbeddedResource Include="MasterForm.resx">
+      <DependentUpon>MasterForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
